+++ lib/glimmer/dsl/xml/xml_text_expression.rb
@@ -1,0 +1,22 @@
+require 'glimmer/dsl/expression'
+require 'glimmer/xml/node'
+
+module Glimmer
+  module DSL
+    module XML
+      class XmlTextExpression < Expression
+        def can_interpret?(parent, keyword, *args, &block)
+          (parent == nil or parent.is_a?(Glimmer::XML::Node)) and
+          (keyword.to_s == "text") and
+          ((args.size == 1) and (args[0].is_a?(String))) and
+          !block
+        end
+
+        def interpret(parent, keyword, *args, &block)
+          parent.children << args[0] if parent
+          args[0]
+        end
+      end
+    end
+  end
+end