+++ lib/glimmer/xml/xml_visitor.rb
@@ -49,8 +49,8 @@ module Glimmer
       end
 
       def append_attributes(node)
-        Glimmer.logger.debug "Take 3"
-        Glimmer.logger.debug(node.attributes)
+        Glimmer.logger&.debug "Take 3"
+        Glimmer.logger&.debug(node.attributes)
         node.attributes.each_key do |attribute|
           attribute_name = attribute
           attribute_name = "#{attribute.name_space.name}:#{attribute.name}" if attribute.is_a?(Node)